backend/requirements.txt
gunicorn==20.1.0
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
PyYAML==6.0
django-import-export==4.1.1
django-filter==23.5
drf_extra_fields==3.7.0